Version 0.8.0 Beta.1 [Prerelease]
First beta release of 0.8.0

+++ Core/Systems/_Essentials/MergeBlackboardsSystem.cs
@@ -13,16 +13,15 @@ namespace Latios.Systems
     /// A system which combines entities with the BlackboardEntityData components into the worldBlackboardEntity and sceneBlackboardEntity.
     /// The entities with BlackboardEntityData are destroyed.
     /// </summary>
-    [RequireMatchingQueriesForUpdate]
     [DisableAutoCreation]
     [UpdateInGroup(typeof(LatiosWorldSyncGroup), OrderFirst = true)]
     //[UpdateBefore(typeof(ManagedComponentsReactiveSystemGroup))] // Constrained in ManagedComponentsReactiveSystemGroup.
     public partial class MergeBlackboardsSystem : SubSystem
     {
-        private EntityDataCopyKit m_copyKit;
-        private EntityQuery       m_query;
+        EntityDataCopyKit m_copyKit;
+        EntityQuery       m_query;
 
-        private NativeList<ComponentType> m_sceneTypes;
+        NativeList<ComponentType> m_sceneTypes;
 
         protected override void OnCreate()
         {
@@ -44,46 +43,57 @@ namespace Latios.Systems
 
         protected override void OnUpdate()
         {
-            Entities.WithStoreEntityQueryInField(ref m_query).WithStructuralChanges().ForEach((Entity entity, ref BlackboardEntityData blackboardEntityData) =>
-            {
-                var types        = EntityManager.GetComponentTypes(entity, Unity.Collections.Allocator.TempJob);
-                var targetEntity = blackboardEntityData.blackboardScope == BlackboardScope.World ? worldBlackboardEntity : sceneBlackboardEntity;
-                var sceneTypes   = m_sceneTypes;
+            latiosWorld.CreateNewSceneBlackboardEntity();
 
-                ComponentType errorType = default;
-                bool          error     = false;
-                foreach (var type in types)
+            if (!m_query.IsEmptyIgnoreFilter)
+            {
+                Entities.WithStoreEntityQueryInField(ref m_query).WithStructuralChanges().ForEach((Entity entity, ref BlackboardEntityData blackboardEntityData) =>
                 {
-                    if (type.TypeIndex == ComponentType.ReadWrite<BlackboardEntityData>().TypeIndex)
-                        continue;
-                    bool skip = false;
-                    for (int i = 0; i < sceneTypes.Length; i++)
+                    var types        = EntityManager.GetComponentTypes(entity, Unity.Collections.Allocator.TempJob);
+                    var targetEntity = blackboardEntityData.blackboardScope == BlackboardScope.World ? worldBlackboardEntity : sceneBlackboardEntity;
+                    var sceneTypes   = m_sceneTypes;
+
+                    ComponentType errorType = default;
+                    bool          error     = false;
+                    foreach (var type in types)
                     {
-                        if (sceneTypes[i].TypeIndex == type.TypeIndex)
+                        if (type.TypeIndex == ComponentType.ReadWrite<BlackboardEntityData>().TypeIndex)
+                            continue;
+                        bool skip = false;
+                        for (int i = 0; i < sceneTypes.Length; i++)
                         {
-                            skip = true;
-                            break;
+                            if (sceneTypes[i].TypeIndex == type.TypeIndex)
+                            {
+                                skip = true;
+                                break;
+                            }
+                        }
+                        if (skip)
+                            continue;
+
+                        if (blackboardEntityData.mergeMethod == MergeMethod.Overwrite || !targetEntity.HasComponent(type))
+                            MoveComponent(entity, targetEntity, type);
+                        else if (blackboardEntityData.mergeMethod == MergeMethod.ErrorOnConflict)
+                        {
+                            errorType = type;
+                            error     = true;
                         }
                     }
-                    if (skip)
-                        continue;
-
-                    if (blackboardEntityData.mergeMethod == MergeMethod.Overwrite || !targetEntity.HasComponent(type))
-                        MoveComponent(entity, targetEntity, type);
-                    else if (blackboardEntityData.mergeMethod == MergeMethod.ErrorOnConflict)
+                    types.Dispose();
+                    if (error)
                     {
-                        errorType = type;
-                        error     = true;
+                        throw new InvalidOperationException(
+                            $"Entity {entity} could not copy component {errorType.GetManagedType()} onto {(blackboardEntityData.blackboardScope == BlackboardScope.World ? "world" : "scene")} entity because the component already exists and the MergeMethod was set to ErrorOnConflict.");
                     }
-                }
-                types.Dispose();
-                if (error)
-                {
-                    throw new InvalidOperationException(
-                        $"Entity {entity} could not copy component {errorType.GetManagedType()} onto {(blackboardEntityData.blackboardScope == BlackboardScope.World ? "world" : "scene")} entity because the component already exists and the MergeMethod was set to ErrorOnConflict.");
-                }
-            }).Run();
-            EntityManager.DestroyEntity(m_query);
+                }).Run();
+                EntityManager.DestroyEntity(m_query);
+            }
+
+            if (!sceneBlackboardEntity.HasComponent<DispatchedNewSceneTag>())
+            {
+                latiosWorld.DispatchNewSceneCallbacks();
+                sceneBlackboardEntity.AddComponent<DispatchedNewSceneTag>();
+            }
         }
 
         void MoveComponent(Entity srcEntity, Entity dstEntity, ComponentType type)
@@ -97,60 +107,6 @@ namespace Latios.Systems
                 m_copyKit.CopyData(srcEntity, dstEntity, type);
             }
         }
-
-        /*private EntityQuery group;
-
-           protected override void OnCreate()
-           {
-            group = GetEntityQuery(typeof(BlackboardEntityData));
-           }
-
-           protected override void OnUpdate()
-           {
-            int length = group.CalculateEntityCount();
-            length     = group.CalculateEntityCount();
-            if (length <= 1)
-                return;
-            var array = group.ToEntityArray(Unity.Collections.Allocator.TempJob);
-
-            UnityEngine.Debug.Log("Merging singletons, found entities to merge: " + array.Length);
-            for (int a = 0; a < array.Length; a++)
-            {
-                if (array[a] == SingletonEntity)
-                    continue;
-                Entity   singletonEntity = SingletonEntity;
-                Entity   entity          = array[a];
-                object[] entityArg       = new object[] { entity };
-                var      types           = EntityManager.GetComponentTypes(entity);
-                for (int i = 0; i < types.Length; i++)
-                {
-                    if (EntityManager.HasComponent(SingletonEntity, types[i]))
-                        continue;
-
-                    var type = types[i].GetManagedType();
-                    if (typeof(IComponentData).IsAssignableFrom(type) || typeof(ISystemStateComponentData).IsAssignableFrom(type))
-                    {
-                        var     getter  = typeof(EntityManager).GetMethod("GetComponentData");
-                        var     gGetter = getter.MakeGenericMethod(type);
-                        var     setter  = typeof(EntityManager).GetMethod("AddComponentData");
-                        var     gSetter = setter.MakeGenericMethod(type);
-                        dynamic icd     = gGetter.Invoke(EntityManager, entityArg);
-                        gSetter.Invoke(EntityManager, new object[] { singletonEntity, icd });
-                    }
-                    else if (typeof(ISharedComponentData).IsAssignableFrom(type) || typeof(ISystemStateSharedComponentData).IsAssignableFrom(type))
-                    {
-                        var     getter  = typeof(InitMergeSingletonsSystem).GetMethod("GetSharedComponentData");
-                        var     gGetter = getter.MakeGenericMethod(type);
-                        var     setter  = typeof(InitMergeSingletonsSystem).GetMethod("AddOrSetSharedComponentData");
-                        var     gSetter = setter.MakeGenericMethod(type);
-                        dynamic iscd    = gGetter.Invoke(EntityManager, entityArg);
-                        gSetter.Invoke(EntityManager, new object[] { singletonEntity, iscd });
-                    }
-                }
-                EntityManager.DestroyEntity(entity);
-            }
-            array.Dispose();
-           }*/
     }
 }
